REFACTOR
1) Add DOM-helper function [ emptyTaskComponent ] -> works as a template to return a list item

2) Refactor [ display() ] function -> display now only deal with array of tasks (array of objs) iterate over it and call [ emptyTaskComponent() ]

3) EventListeners for task status update AND drag&drop (reordering) are added while displaying / creation of a task.
